Move to RowCellTable
